LPD-38043 ah man, that's funny, people from BPM and I created the same class at different modules at the same time. I placed the file in the wrong module though
