--- modules/test/playwright/fixtures/localizationSiteSettingsPageTest.ts
@@ -1,18 +1,0 @@
-/**
- * SPDX-FileCopyrightText: (c) 2000 Liferay, Inc. https://liferay.com
- * SPDX-License-Identifier: LGPL-2.1-or-later OR LicenseRef-Liferay-DXP-EULA-2.0.0-2023-06
- */
-
-import {test} from '@playwright/test';
-
-import {LocalizationSiteSettingsPage} from '../pages/configuration-admin-web/LocalizationSiteSettingsPage';
-
-const localizationSiteSettingsPageTest = test.extend<{
-	localizationSiteSettingsPage: LocalizationSiteSettingsPage;
-}>({
-	localizationSiteSettingsPage: async ({page}, use) => {
-		await use(new LocalizationSiteSettingsPage(page));
-	},
-});
-
-export {localizationSiteSettingsPageTest};
